Moved ParserInitialisationException.java
